ansible: move to kubespray v2.14.2
This updates us to kubernetes v1.18.2

We haven't updated kubernetes in a while; and we want new offline
deployments to use a recent version. We want to keep up cadence.

There is no offifically supported migration path from our previous
version of kubespray to this one.

Due to the stateless nature of kubernetes, we recommend setting up a new
cluster with this version and the redeploying the stateless workloads.

Kubespray itself only supports 3 kubernetes versions; so with this
checkout it is not possible to update from t